requirements.txt
anyio==4.4.0
appnope==0.1.4
appscript==1.2.5
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
aspose-pdf-for-python-via-java==23.10
aspose-words==24.6.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.15.0
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
comm==0.2.2
comtypes==1.4.4
cryptography==42.0.8
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
docx2pdf==0.1.8
docxcompose==1.4.0
docxtpl==0.17.0
exceptiongroup==1.2.1
executing==2.0.1
fastjsonschema==2.20.0
fire==0.6.0
fonttools==4.53.0
fqdn==1.5.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
ipykernel==6.29.4
ipython==8.25.0
ipywidgets==8.1.3
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
JPype1==1.4.1
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.1
jupyter_server_terminals==0.5.3
jupyterlab==4.2.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
jupyterlab_widgets==3.0.11
lxml==5.2.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.0.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.2.1
notebook_shim==0.2.4
numpy==2.0.0
opencv-python-headless==4.10.0.84
overrides==7.7.0
packaging==24.1
pandocfilters==1.5.1
parso==0.8.4
pdf2docx==0.5.8
pdfkit==1.0.0
pdfminer.six==20231228
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.2
plum-dispatch==1.7.4
prometheus_client==0.20.0
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.22
Pygments==2.18.0
PyMuPDF==1.24.5
PyMuPDFb==1.24.3
pypandoc==1.13
PyPDF2==3.0.1
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-json-logger==2.0.7
PyYAML==6.0.1
pyzmq==26.0.3
qtconsole==5.5.2
QtPy==2.4.1
referencing==0.35.1
reportlab==4.2.2
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.1
Send2Trash==1.8.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
spire==0.4.2
Spire.Doc==12.4.0
stack-data==0.6.3
termcolor==2.4.0
terminado==0.18.1
tinycss2==1.3.0
tomli==2.0.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
uri-template==1.3.0
urllib3==2.2.2
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.11
